+++ src/app/App.jsx
@@ -1,24 +1,15 @@
 import React from 'react';
 import { Router, Switch, Route } from 'react-router-dom';
 import { createBrowserHistory } from 'history';
-import styled from 'styled-components';
 
-import { Header } from '../containers/Header/Header';
+import Header from '../containers/Header';
 import { ROUTES, MOVIE_TYPE as type } from '../constants';
-import MoviePage from '../containers/MoviePage/MoviePage';
+import MoviePage from '../containers/MoviePage';
+import { Wrapper, AppWrapper } from './styles';
 
 const history = createBrowserHistory();
 
 function App() {
-  const Wrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-`;
-  const AppWrapper = styled.div`
-    align-self: center;
-    width: 66%;
-`;
   return (
     <Wrapper>
       <AppWrapper>
