+++ src/app/styles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+`;
+export const AppWrapper = styled.div`
+    align-self: center;
+    width: 66%;
+`;
