+++ src/components/Circular/Circular.jsx
@@ -2,8 +2,6 @@ import React from 'react';
 
 import CircularProgress from '@material-ui/core/CircularProgress';
 
-import './circular.css';
-
 const Circular = () => (
   <div className="circular-wrapp">
     <CircularProgress className="circular" />
