--- src/components/Circular/circular.css
@@ -1,8 +1,0 @@
-.circular-wrapp {
-    display: flex;
-    justify-content: center;
-    height: 800px;
-}
-.circular {
-    margin-top: 270px;
-}
