+++ src/components/MovieList/MovieListItem/MovieListItem.jsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 
-import ModalContainer from '../../../containers/ModalContainer/ModalContainer';
-import MovieFullDescription from '../../MovieFullDescription/MovieFullDescription';
+import ModalContainer from '../../../containers/ModalContainer';
+import MovieFullDescription from '../../MovieFullDescription';
 import noImg from '../../../static/images/no-img.png';
 import { STATIC_URL as imgUrl } from '../../../constants';
