+++ src/components/MovieList/MovieListItem/MovieListItem.stories.js
@@ -1,4 +1,3 @@
-/* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 
 import { storiesOf } from '@storybook/react';
