+++ src/containers/MoviePage/index.js
@@ -1,0 +1,1 @@
+export { default } from './MoviePage';
